Global search for specified added, casenameHMCTSinternal and CaseManagementCategory now being set as part of CreateClaimSpecCallbackHandler.
Reformated unspec global search to do the same in CreateClaimCallbackHandler

Global search behind featuretoggle
