+++ src/main/java/uk/gov/hmcts/reform/civil/handler/callback/camunda/caseassignment/AssignCaseToUserForSpecHandler.java
@@ -72,7 +72,8 @@ public class AssignCaseToUserForSpecHandler extends CallbackHandler {
         coreCaseUserService.assignCase(caseId, submitterId, organisationId, CaseRole.APPLICANTSOLICITORONESPEC);
         coreCaseUserService.removeCreatorRoleCaseAssignment(caseId, submitterId, organisationId);
 
-        if (toggleService.isSpecGlobalSearchEnabled()) {
+        // This sets the "supplementary_data" value "HmctsServiceId to the Unspec service ID AAA6
+        if (toggleService.isGlobalSearchEnabled()) {
             setSupplementaryData(caseData.getCcdCaseReference());
         }
