+++ src/main/java/uk/gov/hmcts/reform/civil/model/CaseManagementCategoryElement.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.civil.model.defaultjudgment;
+package uk.gov.hmcts.reform.civil.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
